feat: move project evaluator gallery to full-page route
